# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tesla Invest Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ANNOUNCEMENTS
...
components:
  schemas:
...
    Notification:
      type: object
      required: [id, user_id, type, title, message, is_read, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        is_read:
          type: boolean
        created_at:
          type: string
...